pyproject.toml
[project]
name = "myfastmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
]
